+++ lib/widgets.dart
@@ -25,3 +25,4 @@ export 'src/widgets/future/future_optional.dart';
 // Widgets - Search
 export 'src/widgets/search/search_field.dart';
 export 'src/widgets/search_selection_sheet/search_selection_sheet.dart';
+export 'src/widgets/search_filter_sheet/search_filter.dart';